requirements.txt
aniso8601==10.0.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-RESTful==0.3.10
flask-restful-swagger-2==0.35
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pytz==2025.2
PyYAML==6.0.2
safrs==3.1.6
six==1.17.0
SQLAlchemy==2.0.40
swagger_ui_bundle==1.1.0
typing_extensions==4.13.0
Werkzeug==3.1.3
pytest==8.3.5
coverage==7.8.0
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
google-cloud-storage==2.5.0
google-cloud-pubsub==2.13.0
pytest==6.2.5
pytest-cov==2.12.1
safrs==3.0.0
